feat/pauta: get list-all funcionando

+++ src/main/java/com/kleiman/votingspring/data/PautaRepository.java
@@ -4,5 +4,5 @@ import org.springframework.data.repository.CrudRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface PautaRepository extends CrudRepository<Pauta, Integer> {
+public interface PautaRepository extends CrudRepository<Pauta, Long> {
 }
